+++ index.html
@@ -8,8 +8,9 @@
   </head>
   <body>
     <div id="root">
-      <input type="text" placeholder="이메일 또는 전화번호" name="" id="" />
-      <input type="text" placeholder="비밀번호" name="" id="" />
+      <div class="bg-gray-100 min-h-screen flex justify-center">
+        <div id="app" class="max-w-md w-full"></div>
+      </div>
     </div>
     <script type="module" src="/src/main.js"></script>
   </body>
