+++ src/page/NotFoundPage.js
@@ -1,0 +1,7 @@
+export default function NotFound() {
+  return `
+     <main class="p-4">
+        NOT
+      </main>
+    `;
+}
